- Removed several variables that were initialized, but not used. - Fixed memory leak that would occur when exiting Skulltag after using bots.cpp. - Fixed a memory leak that would occur when using maps with colored sectors in server mode. - Fixed various memory leaks that would occur when clicking the "X" to close down a server.
SVN r912 (trunk)

+++ src/sv_commands.cpp
@@ -4792,10 +4792,10 @@ void SERVERCOMMANDS_GiveInventoryNotOverwritingAmount( AActor *pReceiver, AInven
 	if ( pInventory != NULL )
 		pItem->Amount = pInventory->Amount;
 
-	SERVERCOMMANDS_GiveInventory( pReceiver->player - players, pItem );
+	SERVERCOMMANDS_GiveInventory( ULONG( pReceiver->player - players ), pItem );
 	// [BB] The armor display amount has to be updated separately.
 	if( pItem->GetClass()->IsDescendantOf (RUNTIME_CLASS(AArmor)))
-		SERVERCOMMANDS_SetPlayerArmor( pReceiver->player - players );
+		SERVERCOMMANDS_SetPlayerArmor( ULONG( pReceiver->player - players ));
 }
 //*****************************************************************************
 //
